pyproject.toml
[project]
name = "wldetect"
version = "0.1.0"
description = "WordLlama Detect - Fast, accurate language detection using static LLM embeddings"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.5",
    "safetensors>=0.4.5",
    "pyyaml>=6.0.2",
    "pydantic>=2.12.4",
    "tokenizers>=0.22.0", # HuggingFace tokenizers (Rust, no PyTorch)
    "psutil>=7.0.0",
    "ml-dtypes>=0.5.0", # For fp8 quantization support
    "numba>=0.62.1",
]

[project.optional-dependencies]
training = [
    "torch>=2.8.0",
    "transformers>=4.57.1",
    "datasets>=4.0.0",
    "huggingface_hub>=0.34.4",
    "tqdm>=4.66.4",
    "rich>=14.2.0",
    "scikit-learn>=1.7.1",  # For metrics
    "tensorboard>=2.20.0",  # For logging
    "matplotlib>=3.10.7",  # For confusion matrix plots
    "seaborn>=0.13.2",  # For confusion matrix heatmaps
]
cu128 = [
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.57.1",
    "datasets>=4.0.0",
    "huggingface_hub>=0.34.4",
    "tqdm>=4.66.4",
    "rich>=14.2.0",
    "scikit-learn>=1.7.1",
    "tensorboard>=2.20.0",  # For logging
    "matplotlib>=3.10.7",  # For confusion matrix plots
    "seaborn>=0.13.2",  # For confusion matrix heatmaps
]
dev = [
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.15.0",
    "ruff>=0.14.7",
    "pre-commit>=4.5.0",
]

[project.scripts]
wldetect = "wldetect.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wldetect"]

[tool.hatch.build.targets.wheel.force-include]
"src/wldetect/models" = "wldetect/models"

[tool.hatch.build.targets.sdist]
include = [
    "src/wldetect/**/*.py",
    "src/wldetect/models/*.safetensors",
    "src/wldetect/models/*.yaml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src/wldetect"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "extra == 'cu128'" },
]
torchvision = [
    { index = "pytorch-cu128", marker = "extra == 'cu128'" },
]
